Prototype of quick download test using BBR
This diff proposes an implementation of a quick download test
using the BBR flavour of TCP where the download runs by default
for ten seconds, like before, however it may be stopped much
earlier than that if BBR reports that the bandwidth is not
growing anymore. This seems to me a reasonable way to use the
instrumentation provided by BBR.

My first approach was to run the test for around ten RTTs since
I was told that BBR should converge within that number of rounds,
yet observing the code on a local machine where I was using tc
to emulate network conditions, I concluded that using the estimate
of the bandwidth provided by BBR is probably more robust.

While working on this diff, I noticed it was more rational to
have a directory called `bbr` for BBR stuff than to have instead
a directory called `netx` for network extensions.