ndt7: rewrite server loop to avoid ticker
The ticker was confusing. We were using a single goroutine and
hence basically, for RTTs >= MinMeasurementInterval, it did occur
that the code was blocked for I/O for more than one RTT.

This is normal. However, the presence of the ticker led me to
think, somehow magically, that a "tick" event was emitted every
MinMeasurementInterval. This was not correct, because, as I
explained above, the code could be blocked doing I/O and hence
the "tick" was not processed within a MinMeasurementInterval.

To make the code less surprising, just roll out a single loop
where we check for the elapsed time and react accordingly.

I have tested locally that this still emits an event every
MinMeasurementInterval, which is what we should except until
the RTT is greater than that.

Solves one TODO for me, the one in which I was confused by
the code not providing one event every MinMeasurementInterval.