bbr ndt: fix: use the four tuple as fd-cache key
I have seen this in the server logs and horror followed:

```
2018/09/18 14:21:51  info Adding to cache: 4444 => 9
2018/09/18 14:21:51  info Removing from cache: 4444 => 9
```

Using the local port is not super useful, sadly. I did not
consider that when I coded the cache, my bad.

A more unique piece of information is the five tuple, but
since we just use TCP currently, a four tuple is ok.

Will this fix the previously seen issue where sometimes
the cache was returning an invalid file descriptor? This
is not known for sure. However, at least now the code
telling me information about the cache will be more useful.