Stop sharing structures with the client
Having shared code between the client and us will make the
code too convoluted. It's actually much better to have a
specification and make sure they both comply to that. This
has the additional advantage that one reading the code of
the client can understand everything without having to dig
into the server side implementation as well.