Address feedback by @pboothe
He mentioned that the specification for ndt7 should have been in
`/SPEC.md` in the toplevel. However, I believe that we probably
also want to add more specifications in the future (I would like
to see there also a NDT classic specification, a specification of
how we save date on disk, etc). For this reason, I have renamed
the file to make it explicit that the file contains the specification
of the ndt7 protocol, not a whole ndt7 specification.