ndt7: emit access logs on stdout, error logs on stderr
This is consistent with the way in which Apache and Nginx are
typically dockerised [1]. I've chosen to use the standard access
log format because it was easy for me to emit and it has been
around since forever, hence there are many parsers for it.

Regarding the error logs, instead, I've chosen to use the JSON
format offered by github.com/apex/log, since using structured
logging seems to me more easy to parse.

In the event in which one wants to parse a stream composed of
both access and error logs (as in the output produced by the
`docker logs` command), separating access and error logs is a
straightforward task, since error logs are JSON based and
always start with a `{` character.