ndt7: separate measurements from the sending loop
This separation allows us to perform measurements using the
correct ticking interval and to insert such measurements into
the client stream with correct timing. This doesn't mean the
measurements will be sent timely, because of queue dynamics
but at least their timing will be marked correctly and the
client will receive one event for every ticking period (and
possibly every ticking period if there's no big queue).

Comment out the algorithm for stopping early, for now, since
it has not been used in practice. Let's just collect data.

Change the spec so that application level measurements are
now optional. The rationale is that we get timing annotated
measurements from the measurement goroutine, and we can't
just add app level measurements in there, because the time
annotation will be wrong. Opt not to send such measurements
when sending data to the client, since that is anyway not
a super interesting amount of data, and when we'll collect
more TCP_INFO stats we'll have this information anyway.

The goroutine takes a context telling it whether to stop
early and signals its stop to the main loop by closing the
channel. In the even in which we leave the main loop before
the goroutine closes the channel, use the context and then
wait for the goroutine channel to drain before leaving
the download measurement scope.