bbr: reimplement getting stats in pure go
This simplifies the build, because we don't need anymore to
compile a small bit of C to get the `ndt-server`.

Additionally, this makes the interface between the `bbr` module
and the rest of the tree more flexible, since we can fill the
`model.BBRInfo` structure directly in the Go code.