spec: be pedantic and clarify the correct JSON integer bounds
As mentioned in the spec, this is quite an academic concern for
the moment and most likely in the future. However, clarify we have
thought about this issue, we currently consider it relevant for
pedantic implementations only, and recommend pedantic code to
make sure the int53 boundary is correctly respected.

We may change our mind in the future. (The most likely course of
action is that we'll specify what a pedantic implementation is
supposed to do in such case. For now it doesn't matter.)