ndt7/s/r/file.go: include test type in the file name
While updating the documentation, I realised that it was a good
thing to also include the test type in the file name.

Doing that here, since this PR already concerns itself with the
name of the file and we can easily not squash merge, so that
the process with which I constructed this branch is 100% clear.