Start sketching out common download/upload structure
Here the key idea is that both the download and the upload
perform some common tasks:

1. they both need to upgrade the connection in a very
similar fashion

2. they both need to zip results coming both from the
server and from the client and save them

So, all these bits of code could be performed by
common layers of code.

A potential bummer is that this design binds the same
handler to different paths, so we need to have some
ifs to select what to do. A possible alternative design
to employ here is to use a lambda to implement the
handler and inside the lambda call a different method
of the handler depending on the path.