Hammer the download until it looks good (#82)
* ndt7: refactor to isolate download's measurer

This commit moves the download measurer into a separate file. Now
we have a nice API that we can invoke to measure.

This commit is part of a serious of commit that aim to refactor
the download to be a pipeline of goroutines.

* ndt7: move download's sender in a separate file

Continue reorganizing the download as a pipeline.

* measurer.go downloader.go: stop using context

Rather than performing an explicit check, just use the
context to stop running the download.

* ndt7: refactor sender to be an intermediate stage

We need to add a final stage after the sender, which will
be responsible of reading counter flow messages and of
closing the WebSocket channel cleanly.

* ndt7: add download stage reading counter-flow messages

For now the pipeline does not terminate because the final
stage with a maximum runtime is still missing.

* ndt7: implement the close protocol

This implements the close protocol described in #75. It works
for both half-duplexa and full-duplex clients.

* sender.go: timeout now unnecessary

We have a per-test timer that stops us. No need to also set
specific timers for each individual I/O call.

* ndt7: changes after code review

* sender.go: remove redundant code

We already have conditional nonblocking reading in the next
stage, so there is no point in attempting to spam the channel
since the downstream stage will not block.

* download: roll out better design

1. make sure that all components of pipelines provide liveness
   guarantees by using I/O timeouts _and_ context deadlines

2. use the well know golang pattern for zipping channels, which
   is much simpler to understand given liveness guarantees

3. use a toplevel cancellable context for the whole test such
   that we can interrupt all goroutines early in case we cannot
   successfully save the results on disk

4. use internal messages for keeping track of whether a message
   is originating from the server or from the client

5. remove the ad-hoc pipeline where we were propagating errors
   thus leading to a less surprising design

6. use warnonerror wherever appropriate

* ndt-server.go: forward -datadir to ndt7

* receiver.go: fix: don't pass nil error to logger

* ndt7/client.go: βασιλεὺς βασιλέων, βασιλεύων βασιλευόντων