Start refactoring client and server code (#95)
Related to #22. We want to start detaching the server codebase
and the client we're using for testing.

Ideally, the client we'll use for testing will not event be
this one, but the one we're working on in a separate repo.

One step at a time, tho. Let's first move as much of the client
code as possible outside of the `./ndt7` subtree and use the
minimal amount of dependencies to implement such client.

This change is implemented such that everything should continue
to work as before. Yet, it enables to flatten a little bit server
code and basically assume anything under `./ndt7` is server
side code plus some structure definitions.

More refactoring will be enabled by this refactoring in pure
"yak shaving" spirit.