Factor out common code for saving results
As explained in the documentation of the new package, `saver`, we're
factoring out this common piece of code because we really know it's
gonna be needed by both the download and the upload.

My goal with all this yak shaving is to parachute us safely from
a state where the download works to a state in which both the
download and the upload work and they look like correct.

More changes to come.