ndt7-protocol.md: clients mlab-ns requirements
This diff modifies the ndt7 specification to introduce client
requirements related to mlab-ns.

I have added more information, at the cost of possibly duplicate
some information and/or being redundant with mlab-ns docs, because
IMO this helps us be sure that people who are writing clients for
ndt7 understand how to interact with the platform.

Other solutions to the same problem may be possible and I don't
have a strong opinion that this is the best way.

Closes #34.