ndt7/measurer.go: simplify loop
Rather than looping forever and using select, we can check for the
context being done in the for condition, and do the rest of the work
inside of the loop body. That's less lines of code.

The implementation does not change. We'll poll TCP_INFO/TCP_CC_INFO
every 0.25 seconds and post on the channel, then check whether we
are done just like we did before this diff.