ndt7: improve usage of deadlines
Now that we've fixed the place where we set deadlines, let's turn our
attention onto the values of the deadlines.

It does not make any sense to set a long deadline in the closer where
instead we aim to cut it short at this point of the experiment.

For all senders, the correct I/O deadline is 13 seconds rather than
10 seconds, because that gives some leeway for closing the connection
in a clean way when everything is good.