ndt7: always include ConnectionInfo
We noticed how annoying is not doing that with @robertodauria during
the review of https://github.com/m-lab/ndt7-client-go/pull/39.

As part of such work, we also double checked that every TCPInfo
field described by the spec is a meaningful ndt7 summary.

As a result we can now close #194.