fix(docker-compose.yml): make sure it works on macOS
I am not super happy about exposing 80 and 443 because that would break
running docker-compose on a host that enforces privileged ports.

The fixes in this diff are such that I can run the instructions provided
in the README.md and access ndt7 endpoints.

Sadly, I have no way of testing this on Windows.