Update fullstack script
The script previously just verified that the socket file existed.
The file is created by tcp-info ~immediately, but connecting to it
returns a "connection refused" error for the first few milliseconds.

Instead of just checking for the socket's existence, we attempt to
actually connect to it with socat before starting traceroute-caller
and packet-headers.